Added an eventhandler that automatically hides the popup when the user clicks somewhere outside of it
